+++ features/indents/components/indent-alert-dialogs.tsx
@@ -9,7 +9,7 @@ import {
   AlertDialogFooter,
   AlertDialogHeader,
   AlertDialogTitle,
-} from '@/components/ui/alert-dialog';
+} from '@/components/shadcn/alert-dialog';
 import { Loader2 } from 'lucide-react';
 
 interface DeleteIndentDialogProps {
